fix bug with Ids

+++ app/Http/Controllers/OrganizationController.php
@@ -92,11 +92,14 @@ class OrganizationController extends Controller
     public function deleteMembers(Request $request, $id)
     {
         if (isset($request['userIds'])) {
-            $addMembers = $request['userIds'];
-            foreach ($addMembers as $idMember) {
+            $delMembers = $request['userIds'];
+            foreach ($delMembers as $idMember) {
                 $userRole = DB::table('organization_user')
                     ->select("role")
                     ->where('user_id', '=', $idMember)->get();
+                if(($userRole->count()==0)){
+                    continue;
+                }
                 $role = $userRole->toArray()[0]->role;
                 if ($role != 'admin' && $role != 'owner') {
                     DB::table('organization_user')->where('user_id' , $idMember)->where('organization_id',$id)->delete();
